pyproject.toml
[project]
name = "nonebot-plugin-discord-message-bridge"

version = "1.3.7"

description = "Discord - QQ 消息转发插件"
authors = [
    { name = "xxtg666", email = "xxtg666@gmail.com" }
]
dependencies = [
    "nonebot2",
    "nonebot-adapter-onebot",
    "discord-py",
    "nonebot2[fastapi]",
    "httpx"
]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "GPL-3" }
